+++ pytorch_segmentation_models_trainer/custom_callbacks/__init__.py
@@ -1,9 +1,11 @@
 from .image_callbacks import EnhancedImageSegmentationResultCallback
 from .kl_annealing_callback import KLAnnealingCallback
+from .latent_clustering_callback import AutoencoderLatentClusteringCallback
 from .metrics_callbacks import ConfusionMatrixCallback
 from .training_callbacks import EMACallback, FinalMetricsCallback
 
 __all__ = [
+    "AutoencoderLatentClusteringCallback",
     "ConfusionMatrixCallback",
     "EnhancedImageSegmentationResultCallback",
     "EMACallback",